--- src/main/resources/apknames.txt
@@ -1,2 +1,0 @@
-1
-2
